third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-shorthands-style-format-expected.txt

Test that the format of grid shorthands style uses slashes as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test grid-column shorthand
PASS item.style.gridColumn is "1 / auto"
PASS item.style.gridColumn is "1 / 3"
PASS item.style.gridColumn is "1 / span 2"
PASS item.style.gridColumn is "foo / bar"
PASS item.style.gridColumn is "foo / span bar"
PASS item.style.gridColumn is "2 foo / span 3 bar"
Test grid-row shorthand
PASS item.style.gridRow is "1 / auto"
PASS item.style.gridRow is "1 / 3"
PASS item.style.gridRow is "1 / span 2"
PASS item.style.gridRow is "foo / bar"
PASS item.style.gridRow is "foo / span bar"
PASS item.style.gridRow is "2 foo / span 3 bar"
Test grid-area shorthand
PASS item.style.gridArea is "1 / auto / auto / auto"
PASS item.style.gridArea is "1 / 3 / auto / auto"
PASS item.style.gridArea is "1 / span 2 / auto / auto"
PASS item.style.gridArea is "foo / foo / foo / foo"
PASS item.style.gridArea is "foo / bar / foo / bar"
PASS item.style.gridArea is "2 foo / span 3 bar / auto / auto"
PASS item.style.gridArea is "1 / 2 / 3 / auto"
PASS item.style.gridArea is "1 / 3 / 2 / 4"
PASS item.style.gridArea is "1 / span 2 / 1 / span 2"
PASS item.style.gridArea is "foo / bar / baz / qux"
PASS item.style.gridArea is "foo / span bar / baz / span qux"
PASS item.style.gridArea is "2 foo / span 3 bar / 3 baz / span 2 qux"
PASS successfullyParsed is true

TEST COMPLETE
